HPCC-18515 Dafilesrv remote read streaming
Add support for dynamic remote disk read projects to dafilesrv
(no filtering yet), via JSON requests.

Output format in either binary or JSON.
Continuation either from same instance or if unknown handle,
challenge/response, so client can resend original with cursor.
