+++ common/remote/CMakeLists.txt
@@ -59,6 +59,7 @@ include_directories (
          ./../../rtl/eclrtl
          ./../../system/security/securesocket
          ./../../testing/unittests
+         ./../../rtl/include
     )
 
 ADD_DEFINITIONS( -D_USRDLL -DREMOTE_EXPORTS )
@@ -67,6 +68,7 @@ HPCC_ADD_LIBRARY( remote SHARED ${SRCS}  )
 install ( TARGETS remote RUNTIME DESTINATION ${EXEC_DIR} LIBRARY DESTINATION ${LIB_DIR} )
 
 target_link_libraries ( remote 
+    eclrtl
     jlib
     jhtree 
     mp
