+++ esp/services/ws_workunits/ws_workunitsService.cpp
@@ -45,6 +45,8 @@
 #include "fvdatasource.hpp"
 #include "fvresultset.ipp"
 
+#include "rtlformat.hpp"
+
 #include "package.h"
 
 #ifdef _USE_ZLIB
